Tippy: Allow `update options` effect to run on 2nd render instead of 3rd
Normally, we want to skip updating the Tippy instance's options on the
first and second render (which is automatic due to `setMounted`()) because
it is assumed that no options are changed in this timeframe.

However, this assumption isn't entirely correct, as updating options via props
(causing a re-render) before `setMounted()` will result in the option
not actually being updated (due to the skipping optimization).

To fix this, we allow the Tippy instance to update its options on the 2nd render
and afterwards.